Change chunkSize type from int to ssize_t to reduce latency in min(a, b)

+++ src/collectives/device/all_reduce.h
@@ -222,8 +222,8 @@ __device__ void ncclAllReduceLLKernel(struct CollectiveArgs* args) {
   const ssize_t size = args->N;
   //const int rank = comm->rank;
   const int nranks = comm->nRanks;
-  int chunkSize = llSliceSize * sizeof(uint64_t) / sizeof(T);
-  const ssize_t loopSize = args->nRings*nranks*(ssize_t)chunkSize;
+  ssize_t chunkSize = llSliceSize * sizeof(uint64_t) / sizeof(T);
+  const ssize_t loopSize = args->nRings*nranks*chunkSize;
 
   uint64_t step = ring->send.conn.llStep;
   uint32_t pflag, nflag = step + 1;
